upload events, though it's not perfect
We can upload events to Wordpress, and we event set a meta-tag to
uniquely identify events by their Meetup event ID. But for some reason
the categories aren't populating on the WP events; and although I can
view the published events on the WP admin page, I can't view them on the
website. Ah well, we'll get there.